requirements.txt
selenium==3.141.0
#pytest==5.4.2
allure-pytest-bdd==2.8.22
allure-python-commons==2.8.22
pytest_bdd==4.0.1
pytest-xdist